Creating UpdateInfoScreen style file and removing unusable imports
